requirements.txt
asgiref==3.8.1
Django==5.1.6
django-js-asset==3.1.2
django-tinymce==4.1.0
djangorestframework==3.15.2
git-filter-repo==2.47.0
pillow==11.1.0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
